+++ crates/ruma-client-api/CHANGELOG.md
@@ -23,6 +23,8 @@ Improvements:
   parameter. The former was removed in Matrix 1.14.
 - The `authentication` field of `discovery::discover_homeserver::Response` has
   been removed in favour of `discovery::get_authorization_server_metadata`.
+- The `discovery::discover_homeserver::Response` well-known now includes the
+  `rtc_foci` as defined in MSC4143.
 
 # 0.20.2